Add config.py file
